Feat: Adding Options Menu some logic
Options Menu can be accessed from main menu and in game hud.
Refactoring some components in order to receive support opening and closing the Menu Component.
Added two new components, that allow the player exit the current match and go back to main menu,
or opening the menu while the game is paused.
The ContextProvider was moved upper in the component hierarchy in order to react the entire application
allowing to open the Options Menu from the main screen.