Feat: Adding player prompt when setting new highScore and adding Leaderboard
When a new high score is set, a player name will be asked.
Refactoring some components in order to support game prompt the user.
If the prompt is ignored, there will be no player name but the score will still be set to anonymous score.
The Leaderboard can be accessed from the main menu.